requirements.txt
beautifulsoup4==4.12.3
iitkgp_erp_login==2.4.1
pandas==2.0.2
Requests==2.31.0
tabulate==0.9.0
PyJWT
python-dotenv
